requirements.txt
# python==3.11.11

papermill==2.6.0 # for running notebooks from command line

pandas==2.2.3
openpyxl==3.1.5
# numpy==2.2.6 # auto comes with pandas
torch==2.7.0
accelerate==1.7.0
langdetect==1.0.9
urlextract==1.9.0
scipy==1.15.3
scikit_learn==1.6.1

# models for experiments
openai==1.107.1
google-genai==1.37.0
xai_sdk==1.1.0
transformers==4.57.1
# tokenizers==0.22.1 # auto comes with transformers
bitsandbytes==0.48.2
sentence_transformers==5.1.1

# graphs
wordcloud==1.9.4 # this will also install matplotlib etc.
seaborn==0.13.2

# evals
rouge_score==0.1.2 # this will auto intall nltk etc.
bert_score==0.3.13